backend/requirements.txt
# TradingBot Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Exchange Integration
ccxt>=4.2.0

# Configuration
pyyaml>=6.0.1

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async utilities
aiohttp>=3.9.0
asyncio>=3.4.3
websockets>=12.0

# Email
aiosmtplib>=3.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Utilities
python-dateutil>=2.8.0
